requirements.txt
flask==2.3.3
numpy==1.24.3
scikit-learn==1.2.2
joblib==1.2.0
python-dotenv==1.0.0
gunicorn==20.1.0
setuptools==65.5.0
Werkzeug==2.3.7